Refactor MiR connector command and configuration handling
- Updated README.md to reflect the new command name `inorbit_mir_connector`.
- Modified setup.py to change the entry point for the connector.
- Adjusted Dockerfile CMD to use the updated command name.
- Introduced new configuration model in connector_model.py for better structure and validation.
- Refactored connector.py to utilize the new configuration model.
- Removed deprecated log level handling from various components.
- Updated tests to align with the new configuration structure and command changes.
