+++ mir_connector/systemd/mir-connector.service
@@ -9,7 +9,7 @@ User=inorbit
 # By default, the kill signal is SIGTERM so we override it by SIGINT
 # so the connector handles it the same way as a keyboard interrupt
 KillSignal=SIGINT
-ExecStart=/bin/bash -c 'cd ~/mir-connector-<robot_id> && export $(grep -v '^#' config/.env | xargs) && source ./venv/bin/activate && inorbit-mir100-connector -c config.yaml -id <robot_id>'
+ExecStart=/bin/bash -c 'cd ~/mir-connector-<robot_id> && export $(grep -v '^#' config/.env | xargs) && source ./venv/bin/activate && inorbit_mir_connector -c config.yaml -id <robot_id>'
 
 [Install]
 WantedBy=multi-user.target
